Remove unused cast.

+++ metrics-core/src/main/java/com/yammer/metrics/UniformSample.java
@@ -74,7 +74,7 @@ public class UniformSample implements Sample {
 
     @Override
     public Snapshot getSnapshot() {
-        final int s = (int) size();
+        final int s = size();
         final List<Long> copy = new ArrayList<Long>(s);
         for (int i = 0; i < s; i++) {
             copy.add(values.get(i));
